+++ extensions/word-search/src/components/SearchResults.tsx
@@ -8,6 +8,7 @@ import { Vocabulary } from "@/types";
 
 import useOptionalSelection from "@/hooks/use-optional-selection";
 import useSearchWords from "@/hooks/use-searchwords";
+import { useShowDetailsByDefault } from "@/hooks/use-settings";
 
 import Actions from "@/components/Actions";
 import VocabularySwitch from "@/components/VocabularySwitch";
@@ -42,8 +43,10 @@ export default function SearchResults(
   launchProps: LaunchProps,
   { helperTitle, helperDescription, useVocabulary }: extraOptions = {},
 ) {
+  const showDetailsByDefault = useShowDetailsByDefault();
+
   const [search, setSearch] = useState<string>("");
-  const [showDetails, setShowDetails] = useState<boolean>(false);
+  const [showDetails, setShowDetails] = useState<boolean>(showDetailsByDefault);
   const [vocabulary, setVocabulary] = useState<Vocabulary>(Vocabulary.English);
 
   useOptionalSelection(setSearch, typeof launchProps.fallbackText !== "undefined" && launchProps.fallbackText !== "");