+++ extensions/word-search/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Word Search Changelog
 
+## [Default open details] - {PR_MERGE_DATE}
+
+- Added a preference to open details by default ([#14920](https://github.com/raycast/extensions/issues/14920))
+
 ## [Upgrades] - 2024-06-01
 
 - Updated dependencies to latest versions